[Ref/#462] 사용하지 않는 파일 및 코드 정리, 폴더링 수정, 모델명 규칙에 맞게 수정
