+++ ILSANG/Sources/Domain/Entity/User/User.swift
@@ -1,4 +1,3 @@
-import Foundation
 
 struct User {
     let id: String
