+++ ILSANG/Sources/Global/ViewModelItem/QuestViewModelItem.swift
@@ -81,33 +81,6 @@ class QuestViewModelItem: Hashable, Identifiable {
     var coupon: CouponItem? { coupons.first }
     var hasCouponReward: Bool { !coupons.isEmpty }
     
-    func updateChallengeImages() async {
-//        guard let missions else { return }
-        
-        let challengeImageIds = missions
-            .compactMap { $0.exampleImageIds }
-            .flatMap { $0 }
-        
-        let newImages = await withTaskGroup(of: UIImage?.self) { group -> [UIImage] in
-            var images: [UIImage] = []
-            
-            for challengeImageId in challengeImageIds {
-                group.addTask {
-                    await ImageCacheService.shared.loadImageAsync(imageId: challengeImageId)
-                }
-            }
-            
-            for await image in group {
-                if let image = image {
-                    images.append(image)
-                }
-            }
-            
-            return images
-        }
-        self.challengeImages = newImages
-    }
-    
     func totalRewardPoint() -> Int {
         self.rewards?.reduce(0) { $0 + $1.point } ?? 0
     }
