+++ ILSANG/Sources/Manager/SeasonManager.swift
@@ -49,8 +49,8 @@ final class SeasonManager: ObservableObject {
         let result = await seasonNetwork.getSeasons()
         switch result {
         case .success(let res):
-            self.seasons = res
-            self.currentSeason = res.first(where: { $0.containsToday() })
+            self.seasons = res.map { $0.toDomain() }
+            self.currentSeason = seasons.first(where: { $0.containsToday() })
         case .failure(let err):
             Log(err)
         }
