+++ ILSANG/Sources/Network/Response/ChallengeResponse.swift
@@ -1,0 +1,11 @@
+//
+//  ChallengeResponse.swift
+//  ILSANG
+//
+//  Created by Lee Jinhee on 10/4/25.
+//
+
+
+struct ChallengeResponse: Decodable {
+    let resultCode: String
+}
