+++ ILSANG/Sources/Views/MyPage/UserMissionHistory/UserMissionHistoryViewModelItem.swift
@@ -12,9 +12,9 @@ class UserMissionHistoryViewModelItem {
     let missionHistoryId: Int
     let title, createdAt: String
     let submitImageId: String?
-     var submitImage: UIImage?
+    var submitImage: UIImage?
     let questImageId: String?
-     var questImage: UIImage?
+    var questImage: UIImage?
     let viewCount: Int
     var likeCount: Int
